improved error checking by applying classes

+++ domain-core/src/use-cases/roomUsesCases.ts
@@ -1,7 +1,9 @@
 import type { Room } from "../entities/Room";
+import { NotFoundError } from "../errors";
 import { RoomRepository } from "../repository/RoomRepository";
 
 export class RoomUsesCases implements RoomRepository {
+export class RoomUsesCases {
   constructor(private roomRepository: RoomRepository) {}
 
   async getAll(): Promise<Room[]> {
@@ -14,21 +16,27 @@ export class RoomUsesCases implements RoomRepository {
     return result;
   }
 
-  async getById(id: string): Promise<Room> {
+  async getById(id: string): Promise<Room | null> {
     const result = await this.roomRepository.getById(id);
+    if(!result) throw new NotFoundError("Room not found")
     return result;
   }
 
-  async save(room: Omit<Room, "id">): Promise<Room> {
-    const result = await this.roomRepository.save(room);
-    return result;
+  async save(room: Omit<Room, "id">): Promise<void> {
+    await this.roomRepository.save(room);
   }
 
   async update(id: string, room: Omit<Room, "id">): Promise<void> {
+    const result = await this.roomRepository.getById(id);
+    if(!result) throw new NotFoundError("Room not found")
+
     await this.roomRepository.update(id, room);
   }
 
   async delete(id: string): Promise<void> {
+    const result = await this.roomRepository.getById(id);
+    if(!result) throw new NotFoundError("Room not found")
+      
     await this.roomRepository.delete(id);
   }
 }
